+++ tasks/src/robocorp/tasks/_constants.py
@@ -1,2 +1,6 @@
 # TODO: Support array and object.
 SUPPORTED_TYPES_IN_SCHEMA = (str, int, float, bool)
+
+DEFAULT_TASK_SEARCH_GLOB = "*task*.py"
+
+MODULE_ENTRY_POINT = "robocorp.tasks"