+++ tasks/tests/tasks_tests/test_collect_tasks_glob/my_task.py
@@ -1,0 +1,6 @@
+from robocorp.tasks import task
+
+
+@task
+def task_on_my_task():
+    pass